third_party/blink/web_tests/http/tests/xmlhttprequest/cache-control-request-expected.txt

Tests that XHR respects "Cache-Control" header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS MemoryCache is not used.
PASS successfullyParsed is true

TEST COMPLETE
